backend/app/requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==3.2.16
django-cors-headers==3.13.0
djangorestframework==3.14.0
gTTS==2.3.1
idna==3.4
install==1.3.5
python-dotenv==0.21.0
pyttsx3==2.90
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
gunicorn==20.1.0
